price of appointment is set accordingly
