+++ lotus_backend/src/main/java/isamrs/tim17/lotus/controller/AppointmentController.java
@@ -294,18 +294,18 @@ public class AppointmentController {
 		app.setStatus(status);
 		app.setStartDate(startDate);
 		app.setEndDate(endDate);
-		app.setAppointmentType(doctor.getSpecialty());
+		app.setAppointmentType(doctor.getSpecialty().getType());
+		app.setPrice(doctor.getSpecialty().getPrice());
 		app = service.save(app);
-		app.setPrice(app.getAppointmentType().getPrice());
 		
 		String contentPatient = "Hello " + patient.getName() + " " + patient.getSurname() + "!\nIn response to your appointment request, we have created a term for you in our centre.\n"
 				+ "The appointment is scheduled for " + startDate + " in room " + room.getName() + ".\n"
-				+ "The doctor's name is " + doctor.getName() + " " + doctor.getSurname() + "and the appointment type is " + doctor.getSpecialty().getName() + ".\n"
+				+ "The doctor's name is " + doctor.getName() + " " + doctor.getSurname() + "and the appointment type is " + doctor.getSpecialty().getType().getName() + ".\n"
 				+ "We look forward to seeing you.\nLotus Clinic Staff";
 		
 		String contentDoctor = "Hello " + doctor.getName() + " " + doctor.getSurname() + "!\nYou have a new appointment.\n"
 				+ "The appointment is scheduled for " + startDate + " in room " + room.getName() + ".\n"
-				+ "The patient's name is " + patient.getName() + " " + patient.getSurname() + " and the appointment type is " + doctor.getSpecialty().getName()+ ".\n"
+				+ "The patient's name is " + patient.getName() + " " + patient.getSurname() + " and the appointment type is " + doctor.getSpecialty().getType().getName()+ ".\n"
 				+ "Lotus Clinic Staff";
 		
 		mailSender.sendMsg(patient.getUsername(), "Appointment notiffication", contentPatient);
