+++ lotus_backend/src/main/java/isamrs/tim17/lotus/model/AppointmentPrice.java
@@ -1,5 +1,9 @@
 package isamrs.tim17.lotus.model;
 
+import java.util.HashSet;
+import java.util.Set;
+
+import javax.persistence.CascadeType;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
@@ -7,6 +11,7 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.ManyToOne;
+import javax.persistence.OneToMany;
 import javax.persistence.Table;
 
 @Entity
@@ -21,6 +26,8 @@ public class AppointmentPrice {
 	private AppointmentType type;
 	@ManyToOne(fetch = FetchType.EAGER)
 	private Clinic clinic;
+	@OneToMany(mappedBy = "specialty", fetch = FetchType.LAZY, cascade = CascadeType.ALL)
+	private Set<Doctor> doctors = new HashSet<Doctor>();
 
 	public long getId() {
 		return id;
@@ -60,6 +67,14 @@ public class AppointmentPrice {
 
 	public void setClinic(Clinic clinic) {
 		this.clinic = clinic;
+	}
+
+	public Set<Doctor> getDoctors() {
+		return doctors;
+	}
+
+	public void setDoctors(Set<Doctor> doctors) {
+		this.doctors = doctors;
 	}
 
 }
